backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.10.1
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# API documentation
python-multipart==0.0.6

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2

# Database drivers - use these when needed
# psycopg2-binary==2.9.9  # PostgreSQL - install in Docker
# aiosqlite==0.19.0       # SQLite - for testing
